Implement redirect tests. Their default level is 8.
Redirect tests verify that a given URL is redirected
to a specified other one. Previously we only had sticky
actions which can be used to check if a redirect action
shows up in the final results, but not what the effect is.

Redirect tests can also be leveraged to verify pcrs commands.