Rebuild user-manual, developer-manual and faq.
Looks like the code status UNRELEASED isn't properly dealt with everywhere,
but as this build is not going to get released, I think we can live with
it for now.

Also ampersands in URLs weren't properly escaped, but hopefully
none of the broken links slipped into this commit. This needs to
be addressed before the next release anyway.